add colors on ft_printf

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: fkoehler <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 19:11:04 by fkoehler          #+#    #+#             */
-/*   Updated: 2016/03/14 21:34:30 by fkoehler         ###   ########.fr       */
+/*   Updated: 2016/03/31 17:31:28 by fkoehler         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct		s_list
 
 void				ft_putchar(char c);
 void				ft_putchar_fd(char c, int fd);
-void				ft_putnchar(char c, int n);
+void				ft_putnchar(char c, size_t n);
 void				ft_putstr(char const *s);
 void				ft_putnstr(char const *s, int n);
 void				ft_putstr_fd(char const *s, int fd);
